+++ R/accounts.R
@@ -98,7 +98,7 @@ connectApiUser <- function(account = NULL, server = NULL, apiKey = NULL, quiet =
   # write the user info
   registerUserApiKey(serverName = target$name,
                     accountName = account,
-                    userId = userId,
+                    userId = user$id,
                     apiKey = apiKey)
 
   if (!quiet) {
@@ -467,8 +467,17 @@ resolveAccount <- function(account, server = NULL) {
   }
 }
 
-isShinyapps <- function(accountInfo) {
-  identical(accountInfo$server, "shinyapps.io")
+isShinyapps <- function(server) {
+  identical(server, "shinyapps.io")
+}
+
+isRPubs <- function(server) {
+  identical(server, "rpubs.com")
+}
+
+isConnectInfo <- function(accountInfo = NULL, server = NULL) {
+  host <- if (is.null(accountInfo)) server else accountInfo$server
+  !isShinyapps(host) && !isRPubs(host)
 }
 
 stopWithNoAccount <- function() {