+++ tests/testthat/test-bundle.R
@@ -35,6 +35,43 @@ test_that("simple Shiny app bundle includes correct files", {
   expect_identical(list.files(bundleTempDir), c("manifest.json", "packrat", "server.R", "ui.R"))
 })
 
+test_that("bundle directories are recursively enumerated", {
+  targetDir <- tempfile()
+  dir.create(targetDir)
+  on.exit(unlink(targetDir, recursive = TRUE))
+
+  # tree that resembles the case from https://github.com/rstudio/rsconnect/issues/464
+  files <- c(
+      "app.R",
+      "index.htm",
+      "models/abcd/a_b_pt1/a/b/c1/1.RDS",
+      "models/abcd/a_b_pt1/a/b/c1/2.RDS",
+      "models/abcd/a_b_pt1/a/b/c1/3.RDS",
+      "models/abcd/a_b_pt1/a/b/c1/4.RDS",
+      "models/abcd/a_b_pt1/a/b/c1/5.RDS"
+  )
+
+  # Create and write each file.
+  sapply(files, function(file) {
+    content <- c("this is the file named", file)
+    targetFile <- file.path(targetDir, file)
+    dir.create(dirname(targetFile), recursive = TRUE, showWarnings = FALSE)
+    writeLines(content, con = targetFile, sep = "\n")
+  })
+
+  infos <- file.info(file.path(targetDir, files))
+  totalSize <- sum(infos$size)
+  totalFiles <- length(files)
+
+  result <- listBundleFiles(targetDir)
+
+  # Files are included in the list, count, and sizes, not directories.
+  # Paths are enumerated relative to the target directory, not absolute paths.
+  expect_identical(result$contents, files)
+  expect_equal(result$totalSize, totalSize)
+  expect_equal(result$totalFiles, totalFiles)
+})
+
 test_that("simple Shiny app bundle is runnable", {
   skip_on_cran()
   bundleTempDir <- makeShinyBundleTempDir("simple_shiny", "shinyapp-simple",
@@ -251,6 +288,11 @@ test_that("writeManifest: Rmd without a python block doesn't include reticulate 
   expect_equal(manifest$metadata$appmode, "rmd-static")
   expect_equal(manifest$metadata$primary_rmd, "simple.Rmd")
   expect_equal(manifest$python, NULL)
+  # Confirm that we have removed packrat entries from our file listing but
+  # retain entries for other files.
+  filenames <- names(manifest$files)
+  expect_false(any(grepl("^packrat/", filenames, perl = TRUE)), filenames)
+  expect_true(any(grepl("simple.Rmd", filenames, fixed = TRUE)), filenames)
 })
 
 test_that("getPython handles null python by checking RETICULATE_PYTHON", {